% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{find.loglik.gamma}
\alias{find.loglik.gamma}
\title{Log Likelihood Function for Gamma with Unknown Shape Family (Internal)}
\usage{
find.loglik.gamma(mle, y, nn)
}
\arguments{
\item{mle}{vector, maximum likelihood estimator}

\item{y}{vector of positive response values}

\item{nn}{vector of positive integer weights}
}
\value{
numeric value of log likelihood function
}
\description{
Log Likelihood Function for Gamma with Unknown Shape Family (Internal)
}
