% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{find.loglik.gaussian}
\alias{find.loglik.gaussian}
\title{Log Likelihood Function for Gaussian Family with Indentity Link (Internal)}
\usage{
find.loglik.gaussian(mle, y, nn)
}
\arguments{
\item{mle}{vector of maximum likelihood estimator}

\item{y}{vector of numeric response values}

\item{nn}{vector of positive integer weights}
}
\value{
numeric value of log likelihood function
}
\description{
Log Likelihood Function for Gaussian Family with Indentity Link (Internal)
}
