% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{find.loglik.logit}
\alias{find.loglik.logit}
\title{Log Likelihood Function for Binomial Family with Logistic Link (Internal)}
\usage{
find.loglik.logit(mle, y, nn)
}
\arguments{
\item{mle}{vector of maximum likelihood estimator}

\item{y}{vector of binary response values}

\item{nn}{vector of positive integer weights}
}
\value{
numeric value of log likelihood function
}
\description{
Log Likelihood Function for Binomial Family with Logistic Link (Internal)
}
