% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{find.loglik.poisson}
\alias{find.loglik.poisson}
\title{Log Likelihood Function for Poisson Family (Internal)}
\usage{
find.loglik.poisson(mle, y, nn)
}
\arguments{
\item{mle}{vector, maximum likelihood estimator}

\item{y}{vector of positive integer response values}

\item{nn}{vector of positive integer weights}
}
\value{
numeric value of log likelihood function
}
\description{
Log Likelihood Function for Poisson Family (Internal)
}
