% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_main_fns.R
\name{find.mle}
\alias{find.mle}
\title{Exact Algorithm for Isotonic Single-Index Regression Fitting
with 2-D Predictor Space (Internal)}
\usage{
find.mle(x, y, nn, family = "gaussian")
}
\arguments{
\item{x}{matrix of predictor values}

\item{y}{vector of reponse values}

\item{nn}{vector of positive integer weights}

\item{family}{character string naming the error distribution to be used in
the model}
}
\value{
list including alpha index estimate in vector form, maximum
    likelihood estimate of the response in vector form, and single index
    estimate in vector form
}
\description{
Exact Algorithm for Isotonic Single-Index Regression Fitting
with 2-D Predictor Space (Internal)
}
