% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_util_fns.R
\name{find.pMLE}
\alias{find.pMLE}
\title{Find pMLE (Internal)}
\usage{
find.pMLE(alpha, y, x, nn, family)
}
\arguments{
\item{alpha}{vector with magnitude 1, proposed index}

\item{y}{vector of reponse values}

\item{x}{matrix of predictor values}

\item{nn}{vector of positive integer weights}

\item{family}{\code{\link{family}} object passed by \code{simfast}}
}
\value{
list of numeric vectors
}
\description{
Find pMLE (Internal)
}
