% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_util_fns.R
\name{firstrow}
\alias{firstrow}
\title{Give First Row of Matrix, Return if Vector (Internal)}
\usage{
firstrow(x)
}
\arguments{
\item{x}{a matrix or vector}
}
\value{
a vector (first row of the matrix)
}
\description{
Give First Row of Matrix, Return if Vector (Internal)
}
