% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_generic.R
\name{fitted.simfast}
\alias{fitted.simfast}
\title{Extract fitted values of a \code{simfast} object}
\usage{
\method{fitted}{simfast}(object, ...)
}
\arguments{
\item{object}{on object of class \code{simfast}}

\item{...}{other parameters passed to \code{\link{fitted}}}
}
\value{
a numeric vector of fitted values
}
\description{
Extract fitted values of a \code{simfast} object
}
\examples{

# See the example provided in the \code{\link{simfast}} documentation.

}
\author{
Hanna Jankowski \email{hkj@yorku.ca}
    Konstantinos Ntentes \email{kntentes@yorku.ca} (maintainer)
}
