% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_util_fns.R
\name{getkmax}
\alias{getkmax}
\title{Get Elements Greater Than kth Element (Internal)}
\usage{
getkmax(x, k)
}
\arguments{
\item{x}{numeric vector}

\item{k}{positive integer, location index}
}
\value{
numeric vector of index locations
}
\description{
Get Elements Greater Than kth Element (Internal)
}
