% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_generic.R
\name{mat_pred}
\alias{mat_pred}
\title{Predict Values from a \code{simfast} object with a matrix (Internal)}
\usage{
mat_pred(object, data, type, rule, fn, interp, linv)
}
\arguments{
\item{object}{an object of class \code{simfast}}

\item{data}{optional numeric matrix of new data}

\item{type}{character string specifying type of prediction, takes
value \code{'link'} by default for predictions on the scale of
the linear predictors, and takes \code{'response'} for predicted
response values.}

\item{rule}{integer that describes how to handle predictor values
outside the range of the original data. By default, the value
is \code{1}, which uses linear extrapolation to estimate outside
values, but can take value \code{2} which provides the value of
the closest edge point.}

\item{fn}{interpolation function passed from \code{predict.simfast}}

\item{interp}{extrapolation function passed from \code{predict.simfast}}

\item{linv}{inverse link function passed from \code{predict.simfast}}
}
\value{
a numeric vector
}
\description{
Predict Values from a \code{simfast} object with a matrix (Internal)
}
