% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_main_fns.R
\name{search.mle}
\alias{search.mle}
\title{Stochastic Search Algorithm for Isotonic Single-Index Regression Fitting (Internal)}
\usage{
search.mle(
  x,
  y,
  nn,
  family = "gaussian",
  B = 10000,
  k = 100,
  kappa0 = 100,
  tol = 1e-10,
  max.iter = 20,
  print = TRUE
)
}
\arguments{
\item{x}{matrix of predictor values}

\item{y}{vector of reponse values}

\item{nn}{vector of positive integer weights}

\item{family}{character string matching the error distribution and link function}

\item{B}{positive integer, number of index vectors to test}

\item{k}{positive integer, number of alpha tests per iter}

\item{kappa0}{positive integer, kappa0}

\item{tol}{numeric, convergence tolerance}

\item{max.iter}{numeric, maximum number of iterations allowed}

\item{print}{boolean, prints number and tolerance at each iteration step}

\item{doplot}{boolean, plots fit if dim(x)=2}
}
\value{
list including N(>=1) alpha index estimates in matrix form, N maximum
    likelihood estimates of the response in matrix form, and N single index
    estimates in matrix form
}
\description{
Stochastic Search Algorithm for Isotonic Single-Index Regression Fitting (Internal)
}
