% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfast.R
\name{simfast}
\alias{simfast}
\title{Fitting isotonic generalized single-index regression models via maximum likelihood
with formula support}
\usage{
simfast(
  formula,
  data,
  intercept = FALSE,
  weights = NULL,
  family = "gaussian",
  returnmodel = TRUE,
  returndata = TRUE,
  method = "stochastic",
  multialpha = FALSE,
  B = 10000,
  k = 100,
  kappa0 = 100,
  tol = 1e-10,
  max.iter = 20
)
}
\arguments{
\item{formula}{an object of class \code{\link{formula}}, which is a symbolic
description of the model to be fitted. By default, intercepts are NOT
included, so change argument \code{intercept = TRUE} to include one. NOTE: When
including categorical predictors, be sure that you set
\code{options('contrasts')} in your global options to a desired setting. This
function was designed with the \code{'contr.treatment'} option in mind.}

\item{data}{optional data frame (or object coercible by \code{\link{as.data.frame}})
to a data frame) containing the variables in the model. Variables are taken
from \code{environment(formula)} if not found in \code{data}.}

\item{intercept}{optional boolean, if \code{FALSE} (the default value), then the
model given by the formula does not include an intercept value (even when
including a 1, ie. \code{z ~ 1 + x + y} will only include columns for
\code{x} and \code{y}.}

\item{weights}{optional: vector of positive integer weights, with length
\code{n}. Takes default value \code{NULL} which uses equal weights.}

\item{family}{character string naming the error distribution to be used in
the model, \code{'gaussian'} by default. Other options include
\code{'binomial'} (with logit link, equivalent to calling \code{'logit'}).
Other options coming soon.}

\item{returnmodel}{optional boolean that when \code{TRUE} (the default value)
attaches the \code{\link{model.frame}} object to the simfast object.}

\item{returndata}{optional boolean that when \code{TRUE} (the default value)
returns the predictor matrix and response vector in the simfast object.}

\item{method}{when \code{x} has \code{d=2} columns, method can take \code{'exact'}
argument, which uses an exact omptimization method instead of a stochastic
search. If \code{d} does not equal \code{2}, \code{simfast} will give a
warning and automatically continue with a stochastic search (the default
method, \code{method = 'stochastic'}).}

\item{multialpha}{optional boolean, if \code{TRUE}, will return more than one
\code{indexvec} if available (see Value section and Details).}

\item{B}{positive integer, sets number of index vectors to try when maximizing
the likelihood}

\item{k}{positive integer, less than \code{B}}

\item{kappa0}{positive integer, initial value of kappa}

\item{tol}{numeric, sets tolerance for convergence for \code{method = 'stochastic'}.
Will give value of \code{0} if \code{'exact'} is used.}

\item{max.iter}{positive integer limiting number of iterations for
\code{method = 'stochastic'}}
}
\value{
an object of class \code{simfast}.
}
\description{
Fitting isotonic generalized single-index regression models via maximum likelihood
with formula support
}
