% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfast.R
\name{simfast_m}
\alias{simfast_m}
\title{Fitting isotonic generalized single-index regression models via maximum likelihood}
\usage{
simfast_m(
  x,
  y,
  weights = NULL,
  family = "gaussian",
  returndata = TRUE,
  method = "stochastic",
  multialpha = FALSE,
  B = 10000,
  k = 100,
  kappa0 = 100,
  tol = 1e-10,
  max.iter = 20
)
}
\arguments{
\item{x}{numeric \code{matrix} of observed predictor values, with \code{n}
rows and \code{d} columns}

\item{y}{numeric \code{vector} of observed response values, of length \code{n}}

\item{weights}{optional: vector of positive integer weights, with length
\code{n}. Takes default value \code{NULL} which uses equal weights.}

\item{family}{character string naming the error distribution to be used in
the model, \code{'gaussian'} by default. Other options include
\code{'binomial'} (with logit link, equivalent to calling \code{'logit'}).
Other options coming soon.}

\item{returndata}{optional boolean that when \code{TRUE} (the default value)
returns the predictor matrix and response vector in the simfast object.}

\item{method}{when \code{x} has \code{d=2} columns, method can take \code{'exact'}
argument, which uses an exact omptimization method instead of a stochastic
search. If \code{d} does not equal \code{2}, \code{simfast_m} will give a
warning and automatically continue with a stochastic search (the default
method, \code{method = 'stochastic'}).}

\item{multialpha}{optional boolean, if \code{TRUE}, will return more than one
\code{alpha} vector if available (see Value section and Details).}

\item{B}{positive integer, sets number of index vectors to try when maximizing
the likelihood}

\item{k}{positive integer, less than \code{B}}

\item{kappa0}{positive integer, initial value of kappa}

\item{tol}{numeric, sets tolerance for convergence for \code{method = 'stochastic'}.
Will give value of \code{0} if \code{'exact'} is used.}

\item{max.iter}{positive integer limiting number of iterations for
\code{method = 'stochastic'}}
}
\value{
an object of class \code{simfast}.
}
\description{
Fitting isotonic generalized single-index regression models via maximum likelihood
}
