% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{find.loglik}
\alias{find.loglik}
\title{Selects and Runs Likelihood for Given Distribution Family (Internal)}
\usage{
find.loglik(mle, y, nn, family)
}
\arguments{
\item{mle}{vector of maximum likelihood estimator}

\item{y}{vector of numeric response values}

\item{nn}{vector of positive integer weights}

\item{family}{character string matching the error distribution and link function}
}
\value{
numeric value of log likelihood function
}
\description{
Selects and Runs Likelihood for Given Distribution Family (Internal)
}
