% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generic.R
\name{plot.simfast}
\alias{plot.simfast}
\title{Plot the estimated responses and index values of a \code{\link{simfast}} object}
\usage{
\method{plot}{simfast}(x, points = TRUE, weights = TRUE, predictor = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{\link{simfast}}.}

\item{points}{boolean, if \code{TRUE}, plots points on the estimated ridge function.
Plot will only contain a line representing the estimated ridge function if this
option is changed to \code{FALSE}.}

\item{weights}{boolean, if \code{TRUE}, sizes the points according to their scaled weights.
This only occurs if \code{pts = TRUE}, nothing happens otherwise.}

\item{predictor}{boolean, if \code{TRUE}, produces a \code{plotly} scatterplot
of the estimated responses (\code{yhat}) and the observed predictors \code{x}
(instead of \code{indexvals} on the predictor axes). This is an interactive 3-D
scatterplot when \code{x} is 2 dimensional and a regular scatterplot if \code{x}
is one dimensional. \code{plot.simfast} will throw an error with this option
selected if \code{plotly} is not installed or \code{x} is larger than 2-dimensional.}

\item{...}{all other arguments passed to \code{\link{plot}}.}
}
\description{
Plot the estimated responses and index values of a \code{\link{simfast}} object.
Can also be called by running \code{\link{plot}} on a \code{simfast} object.
}
\examples{

# See the example provided in the \code{\link{simfast}} documentation.

}
\author{
Hanna Jankowski \email{hkj@yorku.ca}
    Konstantinos Ntentes \email{kntentes@yorku.ca} (maintainer)
}
