% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfast.R
\name{simfast}
\alias{simfast}
\title{Fitting isotonic generalized single-index regression models via maximum likelihood
with formula support}
\usage{
simfast(
  formula,
  data,
  intercept = FALSE,
  weights = NULL,
  family = "gaussian",
  returnmodel = TRUE,
  returndata = TRUE,
  method = "stochastic",
  multialpha = FALSE,
  B = 10000,
  k = 100,
  kappa0 = 100,
  tol = 1e-10,
  max.iter = 20
)
}
\arguments{
\item{formula}{an object of class \code{\link{formula}}, which is a symbolic
description of the model to be fitted. By default, intercepts are NOT
included, so change argument \code{intercept = TRUE} to include one. NOTE: When
including categorical predictors, be sure that you set
\code{options('contrasts')} in your global options to a desired setting. This
function was designed with the \code{'contr.treatment'} option in mind.
For \code{'binomial'} response, the vector can be binary values or a vector of
proportions, as long as a proper weight vector (a vector of the denominators
of the proportions, see details) is provided. Categorical vectors (character
strings or factors) will automatically be translated into logical vector with
the baseline factor level a 'success' (takes value \code{1}).}

\item{data}{optional data frame (or object coercible to a data frame by
\code{\link{as.data.frame}}) containing the variables in the model. Variables
are taken from \code{environment(formula)} if not found in \code{data}.}

\item{intercept}{optional boolean, if \code{FALSE} (the default value), then the
model given by the formula does not include an intercept value (even when
including a 1, for example: \code{z ~ 1 + x + y} will only include columns for
\code{x} and \code{y}).}

\item{weights}{optional vector of positive integer weights, with length
\code{n}. Takes default value \code{NULL} which uses equal weights.}

\item{family}{a choice of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
Currently supporting any of \code{\link{gaussian}, \link{binomial},
    \link{poisson},} and \code{\link{Gamma}}. The canonical link function is used
by default, but all link functions available for these families are supported.}

\item{returnmodel}{optional boolean that when \code{TRUE} (the default value)
attaches the \code{\link{model.frame}} object to the simfast object. Leave
as \code{TRUE} to properly use the \code{\link{predict}} function.}

\item{returndata}{optional boolean that when \code{TRUE} (the default value)
returns the predictor matrix and response vector in the simfast object.}

\item{method}{when \code{x} has \code{d=2} columns, method can take \code{'exact'}
argument, which uses an exact omptimization method instead of a stochastic
search. If \code{d} does not equal \code{2}, \code{simfast} will give a
warning and automatically continue with a stochastic search (the default
method, \code{method = 'stochastic'}).}

\item{multialpha}{optional boolean, if \code{TRUE}, will return more than one
\code{indexvec} if available (see Value section and Details).}

\item{B}{positive integer, sets number of index vectors to try when maximizing
the likelihood}

\item{k}{positive integer, less than \code{B}}

\item{kappa0}{positive integer, initial value of kappa}

\item{tol}{numeric, sets tolerance for convergence for \code{method = 'stochastic'}.
Will give value of \code{0} if \code{'exact'} is used.}

\item{max.iter}{positive integer limiting number of iterations for
\code{method = 'stochastic'}}
}
\value{
an object of class \code{simfast}, with the following structure:
\describe{
\item{\code{x}}{if \code{returndata = TRUE}, this is the model matrix used to
fit the model, otherwise it is \code{NULL}.}
\item{\code{y}}{if \code{returndata = TRUE}, this is the response vector used to
fit the model, otherwise it is \code{NULL}.}
\item{\code{alphahat}}{\code{alpha} value estimated by the model fit, will return a
matrix of multiple vectors if \code{multialpha = TRUE}}
\item{\code{yhat}}{vector of estimated response values}
\item{\code{indexvals}}{vector of estimated single index values, the matrix product
of \code{x} and \code{alphahat}}
\item{\code{weights}}{vector of the integer weights used in the model fit}
\item{\code{family}}{the \code{\link{family}} function provided to \code{simfast_m}}
\item{\code{link}}{the link function proposed by \code{family}}
\item{\code{tol}}{numeric convergence tolerance acheived during fitting with
\code{method = 'stochastic'}. For \code{method = 'exact'}, this is \code{0}.}
\item{\code{iter}}{number of iterations used to acheieve convergence. For
\code{method = 'exact'}, this is \code{1}.}
\item{\code{method}}{\code{method} used for fitting the model}
\item{\code{model}}{the \code{\link{model.frame}} generated by the formula object
which is used to generate the \code{\link{model.matrix}} and
\code{\link{model.response}} to pass to \code{simfast_m}}
\item{\code{intercept}}{the \code{intercept} rule selected in the argument}
}
}
\description{
Fitting isotonic generalized single-index regression models via maximum likelihood
with support for estimating response values with \code{\link{predict}} and plotting
values with \code{\link{plot}}. Also includes support for formula objects, data frames,
and built-in regression families (see Arguments).
}
\seealso{
\code{\link{simfast_m}} for providing model matrices instead of a formula.
}
\author{
\preformatted{Hanna Jankowski \email{hkj@yorku.ca}
Konstantinos Ntentes \email{kntentes@yorku.ca} (maintainer)
}
}
