% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfast.R
\name{simfast}
\alias{simfast}
\title{Fitting isotonic generalized single-index regression models via maximum likelihood
with formula support}
\usage{
simfast(
  formula,
  data,
  intercept = FALSE,
  weights = NULL,
  offset = NULL,
  family = "gaussian",
  returnmodel = TRUE,
  returndata = TRUE,
  method = "stochastic",
  multiout = FALSE,
  B = 10000,
  k = 100,
  kappa0 = 100,
  tol = 1e-10,
  max.iter = 20
)
}
\arguments{
\item{formula}{an object of class \code{\link{formula}}, which is a symbolic
description of the model to be fitted. By default, intercepts are NOT
included, so change argument \code{intercept = TRUE} to include one. When
including categorical predictors, be sure to set
\code{options('contrasts')} in your global options to a desired setting.
For \code{'binomial'} response, the vector can be binary values or a vector of
proportions, but should include proper weight vector (a vector of the denominators
of the proportions) is provided. Categorical vectors (character strings or
factors) will automatically be translated into a logical vector with
the baseline factor level a 'success' (takes value \code{1}). Poisson responses
can be integer counts or rates, but should include a proper weight vector (a
vector of the denominators of the rates). Offsets can also be specified in the
formula. Note that multiple offsets are combined, and that duplicate
offsets are only counted once.}

\item{data}{optional data frame (or object coercible to a data frame by
\code{\link{as.data.frame}}) containing the variables in the model. Variables
are taken from \code{environment(formula)} if not found in \code{data}.}

\item{intercept}{logical value, if \code{FALSE} (the default value), then the
model given by the formula does not include an intercept value (even when
including a 1, for example: \code{z ~ 1 + x + y} will only include columns for
\code{x} and \code{y}).}

\item{weights}{optional vector of positive integer weights, with length
\code{n}. Takes default value \code{NULL} which uses equal weights.}

\item{offset}{numeric vector of model offsets , with length
\code{n}. Takes default value \code{NULL} which uses no offset.
If an offset is provided here and in the formula, they are combined.}

\item{family}{a choice of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
Currently supporting any of \code{\link{gaussian}, \link{binomial},
    \link{poisson},} and \code{\link{Gamma}}. The canonical link function is used
by default, but all link functions available for these families are supported.}

\item{returnmodel}{logical value that when \code{TRUE} (the default value)
attaches the \code{\link{model.frame}} object to the simfast object. Leave
as \code{TRUE} to properly use the \code{\link{predict}} function.}

\item{returndata}{logical value that when \code{TRUE} (the default value)
returns the predictor matrix and response vector in the simfast object.}

\item{method}{when \code{x} has \code{d=2} columns, method can take \code{'exact'}
argument, which uses an exact optimization method instead of a stochastic
search. If \code{d} does not equal \code{2}, \code{simfast} will give a
warning and automatically continue with a stochastic search (the default
method, \code{method = 'stochastic'}).}

\item{multiout}{logical value, if \code{TRUE}, will return more than one
\code{alpha} vector and \code{yhat} vector if available, separately from the main
estimate (see Value section and Details).}

\item{B}{positive integer, sets number of index vectors to try when maximizing
the likelihood}

\item{k}{positive integer, algorithmic parameter, more info coming, should be less
than \code{B}}

\item{kappa0}{positive integer, initial value of kappa, more info coming}

\item{tol}{numeric, sets tolerance for convergence for \code{method = 'stochastic'}.
Will give value of \code{0} if \code{'exact'} is used.}

\item{max.iter}{positive integer limiting number of iterations for
\code{method = 'stochastic'}}
}
\value{
an object of class \code{simfast}, with the following structure:
\describe{
\item{\code{x}}{if \code{returndata = TRUE}, this is the model matrix used to
fit the model, otherwise it is \code{NULL}.}
\item{\code{y}}{if \code{returndata = TRUE}, this is the response vector used to
fit the model, otherwise it is \code{NULL}.}
\item{\code{alphahat}}{\code{alpha} value estimated by the model fit}
\item{\code{yhat}}{vector of estimated response values}
\item{\code{indexvals}}{vector of estimated single index values, the matrix product
of \code{x} and \code{alphahat}}
\item{\code{weights}}{vector of the integer weights used in the model fit}
\item{\code{family}}{the \code{\link{family}} function provided to \code{simfast_m}}
\item{\code{loglik}}{a numeric value of the log-likelihood at the estimate.}
\item{\code{offset}}{a numeric vector specifying the offset provided in the
model formula.}
\item{\code{tol}}{numeric convergence tolerance acheived during fitting with
\code{method = 'stochastic'}. For \code{method = 'exact'}, this is \code{0}.}
\item{\code{iter}}{number of iterations used to acheieve convergence. For
\code{method = 'exact'}, this is \code{1}.}
\item{\code{method}}{\code{method} used for fitting the model}
\item{\code{model}}{the \code{\link{model.frame}} generated by the formula object
which is used to generate the \code{\link{model.matrix}} and
\code{\link{model.response}} to pass to \code{simfast_m}}
\item{\code{intercept}}{the \code{intercept} rule selected in the argument}
\item{\code{multialphahat}}{returns all estimated \code{alphahat} vectors
if \code{multiout = TRUE} as a matrix if there is more than one, and
as a vector if there is only one.}
\item{\code{multiyhat}}{returns all estimated \code{yhat} vectors
if \code{multiout = TRUE} as a matrix if there is more than one, and
as a vector if there is only one.}
}
}
\description{
Fitting isotonic generalized single-index regression models via maximum likelihood
with support for estimating response values with \code{\link{predict}} and plotting
values with \code{\link{plot}}. Also includes support for formula objects, data frames,
and built-in regression families (see Arguments).
}
\details{
For i=1,...,n, let X_i be the d-dimensional covariates and Y_i be the corresponding
one-dimensional response.   The isotonic single index model is written as

g(mu) = f(a^T x),

where x=(x_1,...,x_d)^T, g is a known link function, a is an d x 1 index vector,
and f is a nondecreasing function.  The algorithm finds the maximum likelihood
estimate of both f and a, assuming that f is an increasing function.  Implementaton
details can be found in ADD REFs, where theoretical justification of our estimator
(i.e. uniform consistency) is also given.   For the identifiability of isotonic
single index models, we refer to REFs.
}
\examples{

## Load esophageal cancer dataset
esoph <- datasets::esoph
str(esoph) # note that three variables are ordered factors
esoph$ntotal <- esoph$ncases + esoph$ncontrols #use as offset

## subset the data frame for training
set.seed(1) # keep from getting data OOB warning in predict()
nobs <- NROW(esoph)
ind <- sample(1:nobs, size = round(nobs * 0.8))
esophtrain <- esoph[ind, ]
esophtest  <- esoph[-ind, ]

## fit a model with formulas, including ordered/regular factors
## and support for offsets. similar syntax to glm()
sfobj <- simfast(ncases ~ offset(log(ntotal)) + tobgp + alcgp + agegp,
                 data = esophtrain, family = poisson(link = 'log'))

glmobj <- glm(ncases ~ offset(log(ntotal)) + tobgp + alcgp + agegp,
              data = esophtrain, family = poisson(link = 'log'))

## Plot the relationship of estimated responses vs. index values
# Not isotonic because of offset
\dontshow{devAskNewPage(ask = FALSE)}
plot(sfobj)
# Y-hats adjusted to same scale
\dontshow{invisible(readline(prompt="Press <Return> to plot the model with rate response values."))}
plot(sfobj, offset = FALSE)

## Predictions from simfast and glm rounded to nearest integer
sfpred <- round(predict(sfobj, newdata = esophtest))
# Note that simfast only predicts 'response' values
sfpred
glmpred <- round(predict(glmobj, newdata = esophtest, type = 'response'))
glmpred

## Compare squared residuals
sum((sfpred - esophtest$ncases)^2)   #simfast prediction
sum((glmpred - esophtest$ncases)^2)  #glm prediction


}
\seealso{
\code{\link{simfast_m}} for providing model matrices instead of a formula, as well
as more examples.
}
\author{
\preformatted{Hanna Jankowski: hkj@yorku.ca
Konstantinos Ntentes: kntentes@yorku.ca (maintainer)
}
}
